% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ticker_stocksplit.R
\name{add_ticker_stocksplit}
\alias{add_ticker_stocksplit}
\title{Add Stock Splits to DB table}
\usage{
add_ticker_stocksplit(ticker, exact = TRUE, ratio, date = lubridate::today())
}
\arguments{
\item{ticker}{search string like "AAPL" or "Henry Hub"}

\item{exact}{either search like or exact string}

\item{ratio}{split ratio (see details)}

\item{date}{date of stock split}
}
\value{
none
}
\description{
Add Stock Splits to DB table
}
\details{
The ratio is a old to new one. 1/3 means, for 1 old stock of 900$ you will get 3 new ones for 300$.

The function aikiaTrade::adjust_ticker_history_stocksplit() needs to be run after the split has been added to
adjust the fin_ticker_history table accordingly.
}
\examples{
\dontrun{
  add_ticker_stocksplit("AMZN", ratio = 1/20, date = '2022-06-06')
}
}
