% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_ticker_history_stocksplit.R
\name{adjust_ticker_history_stocksplit}
\alias{adjust_ticker_history_stocksplit}
\title{Adjust the history table for stock splits}
\usage{
adjust_ticker_history_stocksplit(check_1st = TRUE)
}
\arguments{
\item{check_1st}{if {false} no check will be performed}
}
\value{
none
}
\description{
Adjust the history table for stock splits
}
\details{
Checks if the DB table fin_ticker_stocksplit contains tickers whith a '0' flag
and performs an adjustment of the tickers' history in the original history table
}
\examples{
\dontrun{
  adjust_ticker_history_stocksplit()
}
}
