% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_ticker_watchlist.R
\name{adjust_ticker_watchlist}
\alias{adjust_ticker_watchlist}
\title{Add and Remove Ticker from Watchlist table}
\usage{
adjust_ticker_watchlist(add = NULL, remove = NULL)
}
\arguments{
\item{add}{provide ticker_yh symbols}

\item{remove}{provide ticker_yh symbols}
}
\value{
no return
}
\description{
Add and Remove Ticker from Watchlist table
}
\examples{
\dontrun{
 adjust_ticker_watchlist(add = c("AAPL","ASML","MSFT"))
}
}
