% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_aikia.R
\name{aikia_main}
\alias{aikia_main}
\alias{aikia_main_light}
\alias{aikia_secondary}
\alias{aikia_palette_main}
\alias{aikia_palette_eight}
\alias{aikia_palette_light}
\title{aikia colours}
\usage{
aikia_main()

aikia_main_light()

aikia_secondary()

aikia_palette_main()

aikia_palette_eight()

aikia_palette_light()
}
\value{
The functions return the hexcodes of colours
}
\description{
Collection of aikia colors
}
\examples{
aikia_main()
aikia_main_light()
aikia_secondary()
aikia_palette_main()
aikia_palette_eight()
aikia_palette_light()
}
