% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_db.R
\name{connect_to_db}
\alias{connect_to_db}
\title{Establishing a connection to the aikia database}
\usage{
connect_to_db(database = "fin_data", user = NULL, password = NULL)
}
\arguments{
\item{database}{The database name}

\item{user}{Your MariaDB Username}

\item{password}{Your password}
}
\value{
A DBI connection object
}
\description{
Establishing a connection to the aikia database
}
\examples{
\dontrun{
connect_to_db(user = "testuser", password = "password")
}
}
