% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bb_ticker_hist_data.R
\name{get_bb_ticker_hist_data}
\alias{get_bb_ticker_hist_data}
\title{Update ticker history from Bloomberg}
\usage{
get_bb_ticker_hist_data(
  stock,
  start = NULL,
  end = aikia::val_date(),
  check_max_date_end = FALSE,
  check_min_date_end = FALSE,
  check_min_date_start = FALSE,
  check_max_date_start = FALSE,
  change_ticker = FALSE
)
}
\arguments{
\item{stock}{a bloomberg ticker}

\item{start}{a start date}

\item{end}{an end date}

\item{check_max_date_end}{\code{TRUE} checks the maximum available date in DB nd resets "END"}

\item{check_min_date_end}{\code{TRUE} checks the minimum available date in DB and resets "END"}

\item{check_min_date_start}{\code{TRUE} checks the minimum available date in DB and resets "START"}

\item{check_max_date_start}{\code{TRUE} checks the maximum available date in DB and resets "START"}

\item{change_ticker}{switch ticker from bloomberg to yahoo (needs to be in meta data)}
}
\value{
a tibble
}
\description{
Update ticker history from Bloomberg
}
\examples{
\dontrun{
get_stock_hist_data_bb("AAPL UW", start = "2022-01-01", end = aikia::val_date())
}
}
