% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fred_infos.R
\name{get_fred_infos}
\alias{get_fred_infos}
\title{Get ECO TS infos from Meta Data Table}
\usage{
get_fred_infos(search_term = NULL, exact = FALSE, list_all = FALSE)
}
\arguments{
\item{search_term}{searches in {name*} and {comment} cols}

\item{exact}{if {TRUE} filters for exact matches}

\item{list_all}{if {TRUE} retuens all available series}
}
\value{
a tibble
}
\description{
Get ECO TS infos from Meta Data Table
}
\examples{
\dontrun{
get_fred_infos("oil")
}
}
