% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prtlabs_sensis.R
\name{get_prtlabs_sensis}
\alias{get_prtlabs_sensis}
\title{Get Ticker Sensis from Portfoliolab}
\usage{
get_prtlabs_sensis(symbols = NULL, allocation = NULL)
}
\arguments{
\item{symbols}{Yahoo ticker symbol(s)}

\item{allocation}{weight of ticker(s) in theroretical prt}
}
\value{
a tibble
}
\description{
Get Ticker Sensis from Portfoliolab
}
\details{
Retrieve Ticker and Portfolio wise Sensis like Sharp Ratio, Sortino Ratio and Drawdowns from the Portfoliolabs.com website.
}
\examples{
\dontrun{
get_prtlabs_sensis("AAPL",100)
}
}
