% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regression_correlation.R
\name{get_regression_correlation}
\alias{get_regression_correlation}
\title{Calculate Regression or Plot Correlation}
\usage{
get_regression_correlation(
  symbols = c("AAPL", "DCOILWTICO", "T10Y2Y", "VIXCLS"),
  start_date = lubridate::today() - 180,
  end_date = lubridate::today() - 1,
  prediction = "AAPL",
  plot_correlation = FALSE,
  incl_cor_histo = TRUE,
  ticker_vs_eco = 0,
  verbose = TRUE
)
}
\arguments{
\item{symbols}{provide {ticker} or {eco} symbols}

\item{start_date}{default value today -180}

\item{end_date}{default vlaue today -1}

\item{prediction}{the symbol to predict the regression for}

\item{plot_correlation}{no regression but plot correlation}

\item{incl_cor_histo}{include histogram in correlation plot}

\item{ticker_vs_eco}{offset returns to see lagging effects}

\item{verbose}{retrieve instant feedback}
}
\value{
a Plot
}
\description{
Calculate Regression or Plot Correlation
}
\examples{
\dontrun{
 get_regression_correlation(symbols = c("AAPL","ASML"))
}
}
