% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_te_infos.R
\name{get_te_infos}
\alias{get_te_infos}
\title{Get economic variable infos from trading economic}
\usage{
get_te_infos(
  search_term = NULL,
  country = "all",
  category = "all",
  exact = FALSE,
  list_all = FALSE
)
}
\arguments{
\item{search_term}{searches in {type} col}

\item{country}{select country - default is {all}}

\item{category}{select category - default is {all}}

\item{exact}{if {TRUE} filters for exact matches}

\item{list_all}{if {TRUE} retuens all available series}
}
\value{
a tibble
}
\description{
Get economic variable infos from trading economic
}
\details{
search the trading economic table by country and category for potential matches:
the idea is to find future expectations to currently leading indicators.
}
\examples{
{
get_te_infos("pmi",country = "states",category = "Busi")
}
}
