% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ticker_infos.R
\name{get_ticker_infos}
\alias{get_ticker_infos}
\title{Get ticker infos from Meta Data Table}
\usage{
get_ticker_infos(search_term = NULL, exact = FALSE)
}
\arguments{
\item{search_term}{searches in {name} and {ticker_*}  cols}

\item{exact}{if {TRUE} filters for exact matches}
}
\value{
a tibble
}
\description{
Get ticker infos from Meta Data Table
}
\examples{
\dontrun{
get_ticker_infos("DBK.DE")
}

}
