% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yh_comp_description.R
\name{get_yh_comp_description}
\alias{get_yh_comp_description}
\title{Get Company Descriptions from Yahoo Finance}
\usage{
get_yh_comp_description(symbols = NULL, verbose = FALSE)
}
\arguments{
\item{symbols}{Yahoo ticker symbol(s)}

\item{verbose}{if \code{TRUE} provides instant feedback}
}
\value{
a tibble
}
\description{
Get Company Descriptions from Yahoo Finance
}
\details{
Retrieve Company's profile and operating descriptions from the Yahoo Finance website.
}
