% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yh_financial_statements.R
\name{get_yh_financial_statements}
\alias{get_yh_financial_statements}
\title{Get yahoo company's financial statements}
\usage{
get_yh_financial_statements(
  symbols = NULL,
  period = "quarterly",
  as_pivot_long = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{symbols}{A yahoo ticker symbol}

\item{period}{Income statement: Either annual or quarterly}

\item{as_pivot_long}{if \code{TRUE} returns result as pivot long}

\item{verbose}{if \code{TRUE} provides instant feedback}
}
\value{
a tibble with corporate financials
}
\description{
Get yahoo company's financial statements
}
\examples{
\dontrun{
get_yh_financial_statements("AAPL")
}

}
