% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yh_hist_data.R
\name{get_yh_hist_data}
\alias{get_yh_hist_data}
\title{Update ticker history from Yahoo Finance}
\usage{
get_yh_hist_data(symbols = NULL, verbose = FALSE)
}
\arguments{
\item{symbols}{a Yahoo ticker symbol}

\item{verbose}{if \code{TRUE} provides instant feedback}
}
\value{
a tibble with available price history
}
\description{
Update ticker history from Yahoo Finance
}
\examples{
\dontrun{
get_yh_hist_data("AAPL")
}

}
