% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yh_ticker_by_isin.R
\name{get_yh_ticker_by_isin}
\alias{get_yh_ticker_by_isin}
\title{Match yahoo finance ticker by ISINs}
\usage{
get_yh_ticker_by_isin(q = NULL, verbose = TRUE)
}
\arguments{
\item{q}{ISIN to request for yahoo ticker}

\item{verbose}{if\code{TRUE} provides instant feedback}
}
\value{
A tibble with several company & exchange info
}
\description{
Match yahoo finance ticker by ISINs
}
\examples{
\dontrun{
get_yh_ticker_by_isin("DE0007100000", verbose = TRUE)
}

}
