% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_theme_aikia.R
\name{gt_theme_aikia}
\alias{gt_theme_aikia}
\title{aikia gt theme}
\usage{
gt_theme_aikia(gt_object)
}
\arguments{
\item{gt_object}{a \code{gt}-object}
}
\value{
a customized gt object
}
\description{
aikia gt theme
}
\examples{
\dontrun{
dplyr::tibble(a = c(1:4), b = c(5:8)) \%>\%
gt::gt() \%>\%
gt_theme_aikia()
}
}
