% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_var.R
\name{hist_Value_at_Risk}
\alias{hist_Value_at_Risk}
\title{Historic Value at Risk}
\usage{
hist_Value_at_Risk(
  name = c("AAPL", "F"),
  weights = c(0.5, 0.5),
  start_date = lubridate::today() - 365.25 * 3.5,
  end_date = lubridate::today() - 1,
  confidence = 99,
  observation_points = 783,
  half_life = 0.5,
  verbose = TRUE
)
}
\arguments{
\item{name}{a vector of yahoo tickers}

\item{weights}{weights for each stock}

\item{start_date}{historical start date}

\item{end_date}{end date of time series}

\item{confidence}{confidence intervall}

\item{observation_points}{number of historical observations, e.g. 1y = 261}

\item{half_life}{in years. until half of the distribution is filled}

\item{verbose}{provides instant feedback}
}
\value{
list of VaR, Time Series and VaR Plot
}
\description{
Historic Value at Risk
}
\examples{
\dontrun{
hist_Value_at_Risk(name = c("AAPL","F"),weights = c(0.5,0.5))
}
}
