% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_peer_ratio.R
\name{rank_peer_ratio}
\alias{rank_peer_ratio}
\title{Rank Peers of same Sectors by different Ratios}
\usage{
rank_peer_ratio(
  industry_level_prio = "industry_group",
  sector_prio = "software",
  industry_level_sec = NULL,
  sector_sec = NULL,
  country = NULL
)
}
\arguments{
\item{industry_level_prio}{select an industry level}

\item{sector_prio}{select a sector within the industry}

\item{industry_level_sec}{optional select a 2nd industry level}

\item{sector_sec}{optional select a sector within the 2nd industry}

\item{country}{optional select a vector of countries}
}
\value{
tibble of multiple financial ratios
}
\description{
Rank Peers of same Sectors by different Ratios
}
\examples{
\dontrun{
 rank_peer_ratio(industry_level_prio = "bics_level_2_industry_group_name", sector_prio = "Software & Tech Services")
}
}
