% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_report_list_ratios.R
\name{rank_report_list_ratios}
\alias{rank_report_list_ratios}
\title{List All Possible Ratios}
\usage{
rank_report_list_ratios(ratio_rank_object = NULL)
}
\arguments{
\item{ratio_rank_object}{result file of 'rank_peer_ratio' function}
}
\value{
gt object of possible ratios to use
}
\description{
List All Possible Ratios
}
\examples{
\dontrun{
 rank_report_list_ratios(ratio_rank_object = NULL)
}
}
