% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_report_sector_ratio.R
\name{rank_report_sector_ratio}
\alias{rank_report_sector_ratio}
\title{Rank Peers in General OR by Specific Ratio}
\usage{
rank_report_sector_ratio(ratio_rank_object, ratio = NULL)
}
\arguments{
\item{ratio_rank_object}{result file of 'rank_peer_ratio' function}

\item{ratio}{keep {NULL} to Rank in General OR select a Ratio}
}
\value{
gt object of peer rank (by ratios)
}
\description{
Rank Peers in General OR by Specific Ratio
}
\examples{
\dontrun{
 rank_report_sector_ratio(ratio_rank_object = NULL, ratio = NULL)
}
}
