% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_report_ticker_ratio.R
\name{rank_report_ticker_ratio}
\alias{rank_report_ticker_ratio}
\title{Rank Peers in General OR by Specific Ratio}
\usage{
rank_report_ticker_ratio(ratio_rank_object, ex_ticker = NULL)
}
\arguments{
\item{ratio_rank_object}{result file of 'rank_peer_ratio' function}

\item{ex_ticker}{Select a Yahoo ticker to list all Ratios}
}
\value{
gt object of Ticker Rank in all Ratios
}
\description{
Rank Peers in General OR by Specific Ratio
}
\examples{
\dontrun{
 rank_report_ticker_ratio(ratio_rank_object = NULL, ex_ticker = NULL)
}
}
