% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_aikia.R
\name{scale_fill_aikia_four}
\alias{scale_fill_aikia_four}
\alias{scale_fill_aikia_eight}
\alias{scale_fill_aikia_eleven}
\alias{scale_color_aikia_four}
\alias{scale_color_aikia_eight}
\title{aikia scales}
\usage{
scale_fill_aikia_four()

scale_fill_aikia_eight()

scale_fill_aikia_eleven()

scale_color_aikia_four()

scale_color_aikia_eight()
}
\value{
The functions return ggplot scales
}
\description{
Collection of aikia ggplot scale functions
}
\examples{
\dontrun{scale_fill_aikia_four()}

\dontrun{scale_fill_aikia_eight()}
\dontrun{scale_fill_aikia_eleven()}
\dontrun{scale_color_aikia_four()}
\dontrun{scale_color_aikia_eight()}
}
\seealso{
Other aikia themes: 
\code{\link{theme_aikia_light}()},
\code{\link{theme_aikia}()}
}
\concept{aikia themes}
