% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_ticker_comment.R
\name{set_ticker_comment}
\alias{set_ticker_comment}
\title{Add a personal comment to symbols}
\usage{
set_ticker_comment(symbols = NULL, comments = NULL)
}
\arguments{
\item{symbols}{provide (multiple) ticker_yh symbols}

\item{comments}{provide a comment to add. OR keep empty to remove}
}
\value{
no return
}
\description{
Add a personal comment to symbols
}
\examples{
\dontrun{
 set_ticker_comment("AAPL", comment = c("new comment"))
}
}
