% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_new_ticker_in_db.R
\name{setup_new_ticker_in_db}
\alias{setup_new_ticker_in_db}
\title{Completely setup of new ticker in DB's ticker_meta_data, ticker_history and ticker_sensi_history tables}
\usage{
setup_new_ticker_in_db(
  ticker_bb,
  provide_tic_yh = NULL,
  provide_tic_fv = NULL,
  start_history = aikia::val_date(),
  verbose = FALSE
)
}
\arguments{
\item{ticker_bb}{a Bloomberg ticker}

\item{provide_tic_yh}{set ticker yahoo or search by ISIN}

\item{provide_tic_fv}{set ticker finviz for news scraping}

\item{start_history}{start date of new ticker's price history}

\item{verbose}{\code{TRUE} provides instand feedback}
}
\value{
NULL. Saves directly to DB
}
\description{
Completely setup of new ticker in DB's ticker_meta_data, ticker_history and ticker_sensi_history tables
}
\examples{
\dontrun{
setup_new_ticker_in_db("AAPL UW", provide_tic_yh = "AAPL")
}

}
