% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_mc_forecast.R
\name{single_mc_forecast}
\alias{single_mc_forecast}
\title{Title}
\usage{
single_mc_forecast(
  s_ticker,
  df_hist = NULL,
  hist_length = 120,
  no_sim = 1000,
  days_forecast = 40,
  combine_ts = TRUE,
  lambda = 0.1,
  scale = c("price", "norm"),
  coridor = c(0:0.9999)
)
}
\arguments{
\item{s_ticker}{a symbol part of the 'df_hist' tibble}

\item{df_hist}{a price history table}

\item{hist_length}{which history length is used for the params estimation}

\item{no_sim}{number of simulations performed by MC}

\item{days_forecast}{how many (working) days to be forecasted}

\item{combine_ts}{combine the history ts with the forecast ts}

\item{lambda}{Poisson jump intensity (default 0.1)}

\item{scale}{forecast the 'price' or 'norm' it to 1}

\item{coridor}{probability corridor}
}
\value{
a tibble
}
\description{
Title
}
\details{
Perform a Monte Carlo based forecast based on the 'get_any_history' function.
Tickers are saved in the 'symbol' column and last values in the 'price' column.
You also need to have a 'date' column
}
\examples{
\dontrun{
single_mc_forecast("AAPL",df,scale = 'price')
}
}
