% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticker_pe_px_indication.R
\name{ticker_pe_px_indication}
\alias{ticker_pe_px_indication}
\title{A simple Ffuture Price Indication using the estimated eps and PE Ratio}
\usage{
ticker_pe_px_indication(stock = NULL, set_pe = NULL)
}
\arguments{
\item{stock}{A yahoo ticker symbol}

\item{set_pe}{Optionally set an own PE ratio}
}
\value{
a gt table
}
\description{
A simple Ffuture Price Indication using the estimated eps and PE Ratio
}
\examples{
\dontrun{
ticker_pe_px_indication("AAPL")
}

}
