% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_tic_sensis_history_in_db.R
\name{update_tic_sensis_history_in_db}
\alias{update_tic_sensis_history_in_db}
\title{Update Ticker Sensis of different returns and volas history in DB}
\usage{
update_tic_sensis_history_in_db(
  ticker_list = NULL,
  start_date = FALSE,
  check_duplicates = TRUE,
  verbose = FALSE,
  backwards = 0
)
}
\arguments{
\item{ticker_list}{A list of ticker_yh symbol in DB}

\item{start_date}{a 1st valuation date}

\item{check_duplicates}{if \code{TRUE} deletes duplicates in sensi table}

\item{verbose}{if \code{TRUE} provides instant feedback}

\item{backwards}{number of days to calculate history geginning from start date}
}
\value{
directly saved to DB
}
\description{
Update Ticker Sensis of different returns and volas history in DB
}
\examples{
\dontrun{
update_tic_sensis_in_db(ticker_list = c("AAPL","AMZN"),valuation_date = aikia::val_date(), backwards = 10)
}
}
