% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_tic_sensis_in_db.R
\name{update_tic_sensis_in_db}
\alias{update_tic_sensis_in_db}
\title{Update Ticker Sensis in DB}
\usage{
update_tic_sensis_in_db(
  ticker_list = NULL,
  valuation_date = NULL,
  check_duplicates = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{ticker_list}{A list of ticker_yh symbol in DB}

\item{valuation_date}{A Date to calculate Volas & Returns for}

\item{check_duplicates}{if\code{TRUE} deletes duplicates in sensi table}

\item{verbose}{if\code{TRUE} provides instant feedback}
}
\description{
Update Ticker Sensis in DB
}
\examples{
\dontrun{
update_tic_sensis_in_db(ticker_list = c("AAPL","AMZN"),valuation_date = aikia::val_date())
}
}
