% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prtlabs_sensis.R
\name{get_prtlabs_sensis}
\alias{get_prtlabs_sensis}
\title{Get Ticker Sensis from Portfoliolab}
\usage{
get_prtlabs_sensis(symbol = NULL, verbose = FALSE)
}
\arguments{
\item{symbol}{Yahoo ticker symbol}

\item{verbose}{if \code{TRUE} provides instant feedback}
}
\value{
a list
}
\description{
Get Ticker Sensis from Portfoliolab
}
\details{
Retrieve Ticker and Portfolio-wise Sensitivities like Sharp Ratio, Sortino Ratio and Drawdowns from the Portfoliolabs.com website.
}
\examples{
\dontrun{
get_prtlabs_sensis("AAPL")
}
}
